:trollface:

+++ FileSearchEngine/ElasticDatabase.cs
@@ -9,6 +9,9 @@ public static class ElasticDatabase
     private static ElasticsearchClient Client { get; set; } = null!;
     
     private const string ArticleIndex = "article-index";
+
+    // ReSharper disable once ConditionIsAlwaysTrueOrFalseAccordingToNullableAPIContract
+    public static bool Initialized => Client != null;
     
     public static void Initialize(string username, string password, string url, string fingerprint)
     {
@@ -21,6 +24,9 @@ public static class ElasticDatabase
     
     public static async Task ResetDatabase()
     {
+        if(!Initialized)
+            return;
+        
         var exists = await Client.Indices.ExistsAsync(ArticleIndex);
         if(!exists.Exists)
             await Client.Indices.DeleteAsync<Article>(ArticleIndex);
@@ -41,6 +47,9 @@ public static class ElasticDatabase
 
     public static async Task CleanDatabase()
     {
+        if(!Initialized)
+            return;
+        
         await Client.DeleteByQueryAsync<object>(ArticleIndex, x => x
             .Query(q => q.MatchAll(new MatchAllQuery()))
         );
@@ -48,6 +57,9 @@ public static class ElasticDatabase
     
     public static async Task<string?> AddFile(Article article, int databaseId)
     {
+        if(!Initialized)
+            return null;
+        
         var indexExistsResponse = await Client.Indices.ExistsAsync(databaseId.ToString());
         if(indexExistsResponse.Exists)
             return null;
@@ -58,6 +70,9 @@ public static class ElasticDatabase
     
     public static async Task DeleteFile(string databaseId)
     {
+        if(!Initialized)
+            return;
+        
         if(!int.TryParse(databaseId, out var id))
             return;
         await Client.DeleteAsync(new DeleteRequest(ArticleIndex, id));
@@ -65,6 +80,9 @@ public static class ElasticDatabase
     
     public static async Task UpdateFile(Article article)
     {
+        if(!Initialized)
+            return;
+        
         if(!int.TryParse(article.ElasticId, out var id))
             return;
 
@@ -73,6 +91,9 @@ public static class ElasticDatabase
     
     public static async Task<IEnumerable<Article>> SearchFiles(string query, int? resultCount = null)
     { 
+        if(!Initialized)
+            return [];
+        
         var response = await Client.SearchAsync<Article>(s => s 
             .Index(ArticleIndex) 
             .From(0)
@@ -90,6 +111,9 @@ public static class ElasticDatabase
 
     public static async Task<IEnumerable<Article>> SearchFilesKnn(string query, int? resultCount = null)
     {
+        if(!Initialized)
+            return [];
+        
         var vector = Model.GetVector(query);
         var response = await Client.SearchAsync<Article>(s => s
             .Index(ArticleIndex)
